back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "annotated-types==0.7.0",
    "bcrypt==5.0.0",
    "blinker==1.9.0",
    "click==8.3.1",
    "dnspython==2.8.0",
    "dotenv>=0.9.9",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "Flask==3.1.3",
    "greenlet==3.3.2",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "psycopg2-binary>=2.9.11",
    "pyasn1==0.6.2",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "PyJWT==2.11.0",
    "python-jose==3.5.0",
    "redis==7.2.1",
    "rq>=2.7.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "SQLAlchemy==2.0.47",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "Werkzeug==3.1.6",
]
